requirements.txt
git+https://github.com/ClaudeCoulombe/FrenchLefffLemmatizer.git
lxml[html_clean]
datasets
textblob
vaderSentiment
safetensors
wordcloud
tweepy
requests
plotly
folium
streamlit-folium
streamlit==1.25.0
pandas==2.0.3
matplotlib==3.7.1
seaborn==0.12.2
fpdf==1.7.2
nltk==3.8.1
scikit-learn==1.3.0
python-dotenv==1.0.0
torch>=2.1.0
transformers>=4.36.2